hexlet-code/pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["mikheyev9 <mikheyev9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
dj-database-url = "^2.2.0"
python-dotenv = "^1.0.1"
psycopg = {extras = ["binary"], version = "^3.2.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[virtualenvs]
in-project = false
